Complete task: Currently when I merge a task in the UI and I have other tasks in-progress (I am working on this repo, claude god, using claude god), it seems like the server loses track of the currently in-progress tasks.  The behavior is a little weird, because I can see the stream go a little further on, but then later when I check back on it, it loses those new messages and seems (in the UI) to have frozen to the time when I merged the changes.
This is bad because I want to be able to merge in changes at the time they are finished.  Can you investigate why this is happening deeply and fix?

Sometimes this hanging happens without me merging anything, but merging seems to be a pretty sure way of triggering it.. Ultrathink